+++ src/core/startup.js
@@ -68,8 +68,8 @@ export function startup(settingsConfig = {}, navigationConfig) {
             qaMode.create(window, game, goToScreen);
         };
         loadFonts(game, onFontsLoaded);
-        a11y.setup(game.canvas.parentElement);
 
+        a11y.setup(game.canvas.parentElement);
         fullscreen.listenForTap(game.canvas.parentElement, game);
     }
 }