+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 | Version | Description |
 |---------|-------------|
+| | QAMode now occurs when on a test URL. |
 | 1.0.6 | |
 | | Add removeFromAccessibleButtons and getAccessibleButtons to accessible layer. |
 | 1.0.5 | |